requirements.txt
Django==6.0.3
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-dotenv==1.2.2
psycopg2-binary==2.9.11
celery==5.6.2
redis==7.3.0
requests==2.32.5
drf-yasg==1.21.15
flake8==7.3.0